KIX2018-9968: keep table state if only tab is switched

See merge request Softwareentwicklung/KIXng/frontend/app!3726

+++ src/frontend-applications/agent-portal/modules/table/webapp/core/factory/TableFactoryService.ts
@@ -45,15 +45,19 @@ export class TableFactoryService {
         EventService.getInstance().subscribe(ContextEvents.CONTEXT_REMOVED, this.subscriber);
     }
 
-    public deleteContextTables(contextId: string, objectType?: KIXObjectType | string): void {
+    public deleteContextTables(contextId: string, objectType?: KIXObjectType | string, keepState?: boolean): void {
         if (this.contextTableInstances.has(contextId)) {
             this.contextTableInstances.get(contextId).forEach((table) => {
                 if (!objectType) {
                     table.destroy();
-                    table.deleteTableState();
+                    if (!keepState) {
+                        table.deleteTableState();
+                    }
                 } else if (table.getObjectType() === objectType) {
                     table.destroy();
-                    table.deleteTableState();
+                    if (!keepState) {
+                        table.deleteTableState();
+                    }
                 }
             });
             this.contextTableInstances.delete(contextId);
